Various updates to the elastic search repositories
