+++ .vs/config/applicationhost.config
@@ -163,7 +163,7 @@
             </site>
             <site name="Exceptionless.Api" id="2">
                 <application path="/" applicationPool="Clr4IntegratedAppPool">
-                    <virtualDirectory path="/" physicalPath="C:\Projects\Exceptionless\Source\Api" />
+                    <virtualDirectory path="/" physicalPath="C:\Code\Exceptionless\Source\Api" />
                 </application>
                 <bindings>
                     <binding protocol="http" bindingInformation="*:50000:localhost" />
