+++ Source/Api/Controllers/Base/PermissionResult.cs
@@ -49,6 +49,5 @@ namespace Exceptionless.Api.Controllers {
                 StatusCode = HttpStatusCode.UpgradeRequired
             };
         }
-
     }
 }
