+++ Source/Api/Controllers/OrganizationController.cs
@@ -137,8 +137,8 @@ namespace Exceptionless.Api.Controllers {
         [HttpPatch]
         [HttpPut]
         [Route("{id:objectid}")]
-        public override IHttpActionResult Patch(string id, Delta<NewOrganization> changes) {
-            return base.Patch(id, changes);
+        public override Task<IHttpActionResult> PatchAsync(string id, Delta<NewOrganization> changes) {
+            return base.PatchAsync(id, changes);
         }
 
         /// <summary>
