+++ Source/Tests/Exceptionless.Api.Tests.csproj
@@ -176,6 +176,7 @@
     <Compile Include="Plugins\EventParserTests.cs" />
     <Compile Include="Plugins\SummaryDataTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Repositories\ElasticSearchConfigurationTests.cs" />
     <Compile Include="Repositories\ElasticSearchRepositoryTests.cs" />
     <Compile Include="Repositories\StackRepositoryTests.cs" />
     <Compile Include="Repositories\TokenRepositoryTests.cs" />
